fix: gradation on the bottom to make transition better
