+++ src/components/core/navbar/Logo/index.tsx
@@ -1,5 +1,11 @@
+import { useRouter } from "next/navigation";
 import * as S from "./styles.css";
 
 export const Logo = () => {
-  return <div className={S.container}>Kyle</div>;
+  const router = useRouter();
+  return (
+    <div className={S.container} onClick={() => router.push("/hello")}>
+      Kyle
+    </div>
+  );
 };
