+++ src/components/core/navbar/Logo/styles.css.ts
@@ -1,3 +1,5 @@
 import { style } from "@vanilla-extract/css";
 
-export const container = style({});
+export const container = style({
+  cursor: "pointer",
+});
