+++ src/components/home/Story/SecondStory/index.tsx
@@ -1,0 +1,3 @@
+export const SecondStory = () => {
+  return <div>yo</div>;
+};
