--- src/components/home/StoryList/index.tsx
@@ -1,9 +1,0 @@
-import { FirstStory } from "../Story/FirstStory";
-
-export const StoryList = () => {
-  return (
-    <>
-      <FirstStory />
-    </>
-  );
-};
